Add a mode to the environments where for the first lap, we only disclose the next checkpoint coordinates.
This makes the envs align w/ the games, where only the next checkpoint coords are given. Since in game, you can write the python program to remember all the checkpoints that had been provided so far, after the first lap, it is possible to get the next 2 checkpoint coordinates; therefore in order for the envs to be aligned w/ the game, we withhold the information about the second checkpoint out in the first lap.